다른 아이디의 댓글을 삭제할 수 있는 문제 수정

+++ controller/posts.js
@@ -75,6 +75,13 @@ async function revisePost(req, res) {
 
 async function deletePost(req, res) {
   const { postsId } = req.params;
+  const { nickname } = res.locals.user;
+  const { Writer } = await Posts.findOne({ postsId });
+  console.log(nickname);
+  console.log(Writer);
+  if (nickname !== Writer) {
+    return res.send("Fuck You");
+  }
 
   await Posts.deleteOne({ postsId: postsId });
